Fix incremental build of Material.Blazor project
It should not be necessary to disable Visual Studio's fast up-to-date check. Instead, we just need to configure it correctly when working with additional files during build.

In this case, we add `UpToDateCheckInput` and `UpToDateCheckOutput` for all scripts and styles. Because each of these are compiled individually (that is, changes to scripts don't effect styles, etc) we model them with `Set` metadata that enforces partitioning between these groups of inputs/outputs.

Another change is to remove the `wwwroot\intermediate.js` file from the project. This file is created during build, and deleted as the build completes. If it is not removed, then VS watches for file system activity on this file and considers it's creation and deletion as an indication that the project has changed in some way that requires a further build. It's idiomatic in .NET to generate such temporary files in the intermediate (`obj`) folder, though that pattern is not common in the web world, so on balance it's easier to have the project ignore the file.

With these changes, builds in VS are much faster when no C# code is changed, as the DLL is not regenerated. If no C#/TS/SCSS files have changed, the build completes in milliseconds.

I hope this makes you more productive!